Update ProceduralPartsAllPartsFromStart_MM.cfg
This patch unlocks all PP shapes and sizes from the start.
1) Unlocks all shapes and sizes of procedural parts and
2) Makes all procedural parts available from the start
I believe this, or a variation of this, was already available on PP's home page, this just has a few tweaks thrown in.
This file as well as the next two I've used in tandem with all of my games because I haven't taken to the time to either combine them or figure out which one does the job so I created a couple extra with patch placement changes (which again, can be condensed into one MM patch) as redundancies while in a hurry. I may update this and remove the two below this but as it stands this is how I'm currently using it.